Fix remaining MSVC C4244 warnings in tensor.h and shape_inference.cc
Two root causes triggering C4244 across multiple compilation units:

1. onnx/common/tensor.h: std::accumulate called with int literal '1'
   as initial value but std::multiplies<int64_t> as operator. MSVC
   narrows the int64_t result back to int (_Ty). Fix by using int64_t{1}
   as the initial value, matching the multiplies type. This was triggering
   C4244 in every TU that included tensor.h (ir_pb_converter.cc,
   tensor_util.cc, inliner.cc, convert.cc, helper.cc).

2. onnx/defs/shape_inference.cc line 535: dims(0) returns int64_t
   but getAttributeProtoElemTypeAndLength returns pair<int,int>,
   causing implicit int64_t->int narrowing. Fix with static_cast<int>.

+++ onnx/defs/shape_inference.cc
@@ -532,7 +532,7 @@ std::pair<int, int> getAttributeProtoElemTypeAndLength(const AttributeProto* att
       fail_type_inference(
           "Attribute ", attr_proto->name(), " expected to be a 1D tensor but was ", attr_proto->t().dims_size(), "D");
     }
-    return {attr_proto->t().data_type(), attr_proto->t().dims(0)};
+    return {attr_proto->t().data_type(), static_cast<int>(attr_proto->t().dims(0))};
   }
   return {TensorProto::UNDEFINED, 0};
 }
